feat(EPMUII-10524): CRA to VITE
