+++ package.json
@@ -19,10 +19,13 @@
   "homepage": "https://mriviewer.opensource.epam.com/",
   "private": false,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --detectOpenHandles",
-    "debugtest": "DEBUG=true react-scripts test",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest",
+    "test:ui": "vitest --ui",
+    "test:watch": "vitest --watch",
+    "debugtest": "cross-env DEBUG=true vitest",
     "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
     "format": "prettier --write \"src/**/*.{js,jsx,html,css}\"",
     "docker": "npm run build && docker run --name mri-viewer -v ./build:/usr/share/nginx/html:ro -d -p 8080:80 nginx",
@@ -46,7 +49,6 @@
     "react-modal": "^3.14.2",
     "react-popper": "^2.2.5",
     "react-redux": "7.2.4",
-    "react-scripts": "4.0.3",
     "redux": "4.1.0",
     "seedrandom": "^2.4.4",
     "three": "0.101.1"
@@ -54,24 +56,33 @@
   "devDependencies": {
     "@babel/core": "7.14.3",
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-transform-react-jsx": "^7.25.9",
     "@babel/plugin-transform-runtime": "^7.22.15",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.0",
+    "@babel/preset-react": "^7.26.3",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^12.1.5",
     "@testing-library/react-hooks": "^7.0.2",
     "@testing-library/user-event": "^13.2.1",
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
     "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react": "^4.3.4",
+    "cross-env": "^7.0.3",
     "eslint": "^7.27.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "7.x.x",
     "husky": "^7.0.0",
+    "jsdom": "^26.0.0",
     "lint-staged": "^11.0.0",
     "prettier": "^2.4.1",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "vite": "^6.2.2",
+    "vite-plugin-glsl": "^1.3.3",
+    "vitest": "^3.0.9"
   },
+  "type": "module",
   "jest": {
     "resetMocks": false,
     "coverageReporters": [
