+++ src/engine/lib/core/loaders/MRIFileLoader.ts
@@ -99,8 +99,8 @@ export class MRIFileLoader {
     this.filesLength = fileNames.length;
     const filePromises = fileNames.map((filename: string) => this.fetchSingleFile(base + filename));
     const files = await Promise.all(filePromises);
-    const validFalies = files.filter(Boolean) as Array<File[]>;
-    return validFalies.flat();
+    const validFiles = files.filter(Boolean) as Array<File[]>;
+    return validFiles.flat();
   }
 
   /**
@@ -133,9 +133,9 @@ export class MRIFileLoader {
 
     const files = await Promise.all(filePromises);
 
-    const validFalies = files.filter(Boolean) as Array<File[]>;
+    const validFiles = files.filter(Boolean) as Array<File[]>;
 
-    return validFalies.flat();
+    return validFiles.flat();
   }
 
   /**
