+++ src/ui/FileReaders/OpenFromURLComponent.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { UIButton } from '../Button/Button';
 import ModalUrl from '../Modals/ModalUrl';
-import MriViwer from '../../engine/lib/MRIViewer';
+import MriViewer from '../../engine/lib/MRIViewer';
 
 export const OpenFromURLComponent = ({ cx }) => {
   const [showModalUrl, setShowModalUrl] = useState(false);
@@ -16,7 +16,7 @@ export const OpenFromURLComponent = ({ cx }) => {
 
   const onClickLoadUrl = (url) => {
     setShowModalUrl(false);
-    MriViwer.read(url);
+    MriViewer.read(url);
   };
 
   return (
